- Made separate controllers for User and Event Photos - Changed paths in UI services
