+++ EventsExpress/ClientApp/src/services/PhotoService.js
@@ -3,17 +3,17 @@ import EventsExpressService from './EventsExpressService';
 const baseService = new EventsExpressService();
 
 export default class PhotoService {
-    getPreviewEventPhoto = id => baseService.getPhoto(`photo/GetPreviewEventPhoto/${id}`);
+    getPreviewEventPhoto = id => baseService.getPhoto(`EventPhoto/GetPreviewEventPhoto/${id}`);
 
-    getFullEventPhoto = id => baseService.getPhoto(`photo/GetFullEventPhoto/${id}`);
+    getFullEventPhoto = id => baseService.getPhoto(`EventPhoto/GetFullEventPhoto/${id}`);
 
-    getUserPhoto = id => baseService.getPhoto(`photo/GetUserPhoto/${id}`);
+    getUserPhoto = id => baseService.getPhoto(`UserPhoto/GetUserPhoto/${id}`);
 
     setEventTempPhoto = async (id, data) => {
         let photo = new FormData();
         photo.append(`Photo`, data);
-        return baseService.setResourceWithData(`event/SetEventTempPhoto/${id}`, photo)
+        return baseService.setResourceWithData(`EventPhoto/SetEventTempPhoto/${id}`, photo)
     };
 
-    deleteUserPhoto = id => baseService.setResource(`photo/DeleteUserPhoto/${id}`)
+    deleteUserPhoto = id => baseService.setResource(`UserPhoto/DeleteUserPhoto/${id}`)
 }
