+++ EventsExpress/ClientApp/src/services/UserService.js
@@ -19,7 +19,7 @@ export default class UserService {
     setAvatar = async(data) => {
         let file = new FormData();
         file.append('Photo', data.image.file);
-        return baseService.setResourceWithData(`users/changeAvatar/${data.userId}`, file);
+        return baseService.setResourceWithData(`UserPhoto/changeAvatar/${data.userId}`, file);
     }
 
     setChangeUserRole = data =>
